created update,delete and display with dto

+++ src/main/java/com/example/day3ms/service/StudentService.java
@@ -1,5 +1,8 @@
 package com.example.day3ms.service;
 
+import com.example.day3ms.dto.StudentRequestDto;
+import com.example.day3ms.dto.StudentResponseDto;
+import com.example.day3ms.exception.StudentNotFoundException;
 import com.example.day3ms.model.StudentModel;
 import com.example.day3ms.repository.StudentRepository;
 import org.springframework.stereotype.Service;
@@ -14,27 +17,58 @@ public class StudentService {
         this.repository = repository;
     }
 
-    public StudentModel addStudent(StudentModel student){
-        return repository.save(student);
+//    public StudentModel addStudent(StudentModel student){
+//        return repository.save(student);
+//    }
+    public StudentResponseDto addStudent(StudentRequestDto dto){
+        StudentModel student  = new StudentModel();
+        student.setName(dto.getName());
+        student.setAge(dto.getAge());
+        student.setEmail(dto.getEmail());
+        StudentModel saved = repository.save(student);
+
+        return new StudentResponseDto(
+                saved.getId(),
+                saved.getName(),
+                saved.getAge(),
+                saved.getEmail()
+        );
     }
 
     public List<StudentModel> getStudents(){
         return repository.findAll();
     }
-
+    public List<StudentResponseDto> getAllStudents(){
+        return repository.findAll()
+                .stream()
+                .map(s-> new StudentResponseDto(
+                        s.getId(),
+                        s.getName(),
+                        s.getAge(),
+                        s.getEmail()
+                )).toList();
+    }
     // update
-    public StudentModel updateStudent(String id,StudentModel student){
+    public StudentResponseDto updateStudent(String id,StudentRequestDto dto){
         StudentModel existingStudent = repository.findById(id)
                 .orElseThrow(()->new RuntimeException("No student found"));
 
-        existingStudent.setName(student.getName());
-        existingStudent.setAge(student.getAge());
-        existingStudent.setEmail(student.getEmail());
-
-        return repository.save(existingStudent);
+        existingStudent.setName(dto.getName());
+        existingStudent.setAge(dto.getAge());
+        existingStudent.setEmail(dto.getEmail());
+        StudentModel updated = repository.save(existingStudent);
+        return new StudentResponseDto(
+                updated.getId(),
+                updated.getName(),
+                updated.getAge(),
+                updated.getEmail()
+        );
     }
 
     public void deleteStudent(String id){
+        if(!repository.existsById(id)){
+            throw new StudentNotFoundException("No student found with this id");
+        }
         repository.deleteById(id);
     }
 }
